We all know that SGI's 10x patches are not acceptable as a whole to be included into Apache 1.3. But no one seems to have recognized that Mike Abbott's changes also included very reasonable and good cleanup changes. So I investigated two hours this morning and extracted those non-performance related changes from the 512KB 10xpatch-1.3.11-0 patch and merged them into 1.3. Here is the CHANGES entry:
| Merged in a small subset of SGI's latest `10x' patchkit for Apache
| 1.3.11. The extracted and merged in parts are entirely cleanup and
| non-performance related changes only. SGI's remaining changes are
| not taken over, because they are either cluttering the Apache 1.3
| sources too much (e.g. the lint(1) related changes) or cause too
| much internal changes (e.g. the ap_int32 types, etc.) which are not
| reasonable to do any longer for Apache 1.3 (they should be done for
| Apache 2.0 instead).

Submitted by: Mike Abbott <mja@sgi.com>
Extracted and merged in by: Ralf S. Engelschall
Reviewed by: Ralf S. Engelschall


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@84685 13f79535-47bb-0310-9956-ffa450edef68

+++ src/include/ap_config.h
@@ -1406,7 +1406,7 @@ int gethostname(char *name, int namelen);
 void syslog(int, char *,...);
 char *mktemp(char *);
 
-long vfprintf(FILE *, const char *, va_list);
+int vfprintf(FILE *, const char *, va_list);
 
 #endif /* SUNOS_LIB_PROTOTYPES */
